fix: move hook back
